refactor tasks helpers
